Divide HomeContentFeature, add preview cases
